Separate Board and BoardGroup tests
